Refactoring : Centralized GouvApi calls // Corrected effects on SignIn pages
